+++ src/index.ts
@@ -1,4 +1,7 @@
 import DeviceManager from './device-manager';
 
 export { supportsMediaDevices } from './support-detection';
+export { DeviceInfo, DeviceKind } from './enumerate-devices';
+export { OperationType } from './device-manager';
+
 export default new DeviceManager();